refactor(package): 폴더구조 리팩토링
