+++ src/main/java/lotto/domain/Lottos.java
@@ -4,7 +4,6 @@ import java.util.Collections;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
-import lotto.Lotto;
 
 public class Lottos {
     private final List<Lotto> lottos;
