+++ src/main/java/lotto/service/LottoGenerator.java
@@ -1,6 +1,6 @@
 package lotto.service;
 
-import lotto.Lotto;
+import lotto.domain.Lotto;
 
 // 로또 생성 인터페이스
 @FunctionalInterface
