+++ src/main/java/lotto/service/RandomLottoGenerator.java
@@ -2,7 +2,7 @@ package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.LottoNumber;
 
 public class RandomLottoGenerator implements LottoGenerator {
