+++ src/test/java/lotto/service/RandomLottoGeneratorTest.java
@@ -3,7 +3,7 @@ package lotto.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
